backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
openai
reportlab
pytest


